+++ Source/ExTDSUE/Game/ExTDSUEGameMode.h
@@ -1,0 +1,19 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "ExTDSUEGameMode.generated.h"
+
+UCLASS(minimalapi)
+class AExTDSUEGameMode : public AGameModeBase
+{
+	GENERATED_BODY()
+
+public:
+	AExTDSUEGameMode();
+};
+
+
+
