+++ Source/ExTDSUE/Game/ExTDSUEPlayerController.cpp
@@ -1,0 +1,112 @@
+// Copyright Epic Games, Inc. All Rights Reserved.
+
+#include "ExTDSUEPlayerController.h"
+#include "Blueprint/AIBlueprintHelperLibrary.h"
+#include "Runtime/Engine/Classes/Components/DecalComponent.h"
+#include "HeadMountedDisplayFunctionLibrary.h"
+#include "../Character/ExTDSUECharacter.h"
+#include "Engine/World.h"
+
+AExTDSUEPlayerController::AExTDSUEPlayerController()
+{
+	bShowMouseCursor = true;
+	DefaultMouseCursor = EMouseCursor::Crosshairs;
+}
+
+void AExTDSUEPlayerController::PlayerTick(float DeltaTime)
+{
+	Super::PlayerTick(DeltaTime);
+
+	// keep updating the destination every tick while desired
+	if (bMoveToMouseCursor)
+	{
+		MoveToMouseCursor();
+	}
+}
+
+void AExTDSUEPlayerController::SetupInputComponent()
+{
+	// set up gameplay key bindings
+	Super::SetupInputComponent();
+
+	InputComponent->BindAction("SetDestination", IE_Pressed, this, &AExTDSUEPlayerController::OnSetDestinationPressed);
+	InputComponent->BindAction("SetDestination", IE_Released, this, &AExTDSUEPlayerController::OnSetDestinationReleased);
+
+	// support touch devices 
+	InputComponent->BindTouch(EInputEvent::IE_Pressed, this, &AExTDSUEPlayerController::MoveToTouchLocation);
+	InputComponent->BindTouch(EInputEvent::IE_Repeat, this, &AExTDSUEPlayerController::MoveToTouchLocation);
+
+	InputComponent->BindAction("ResetVR", IE_Pressed, this, &AExTDSUEPlayerController::OnResetVR);
+}
+
+void AExTDSUEPlayerController::OnResetVR()
+{
+	UHeadMountedDisplayFunctionLibrary::ResetOrientationAndPosition();
+}
+
+void AExTDSUEPlayerController::MoveToMouseCursor()
+{
+	if (UHeadMountedDisplayFunctionLibrary::IsHeadMountedDisplayEnabled())
+	{
+		if (AExTDSUECharacter* MyPawn = Cast<AExTDSUECharacter>(GetPawn()))
+		{
+			if (MyPawn->GetCursorToWorld())
+			{
+				UAIBlueprintHelperLibrary::SimpleMoveToLocation(this, MyPawn->GetCursorToWorld()->GetComponentLocation());
+			}
+		}
+	}
+	else
+	{
+		// Trace to see what is under the mouse cursor
+		FHitResult Hit;
+		GetHitResultUnderCursor(ECC_Visibility, false, Hit);
+
+		if (Hit.bBlockingHit)
+		{
+			// We hit something, move there
+			SetNewMoveDestination(Hit.ImpactPoint);
+		}
+	}
+}
+
+void AExTDSUEPlayerController::MoveToTouchLocation(const ETouchIndex::Type FingerIndex, const FVector Location)
+{
+	FVector2D ScreenSpaceLocation(Location);
+
+	// Trace to see what is under the touch location
+	FHitResult HitResult;
+	GetHitResultAtScreenPosition(ScreenSpaceLocation, CurrentClickTraceChannel, true, HitResult);
+	if (HitResult.bBlockingHit)
+	{
+		// We hit something, move there
+		SetNewMoveDestination(HitResult.ImpactPoint);
+	}
+}
+
+void AExTDSUEPlayerController::SetNewMoveDestination(const FVector DestLocation)
+{
+	APawn* const MyPawn = GetPawn();
+	if (MyPawn)
+	{
+		float const Distance = FVector::Dist(DestLocation, MyPawn->GetActorLocation());
+
+		// We need to issue move command only if far enough in order for walk animation to play correctly
+		if ((Distance > 120.0f))
+		{
+			UAIBlueprintHelperLibrary::SimpleMoveToLocation(this, DestLocation);
+		}
+	}
+}
+
+void AExTDSUEPlayerController::OnSetDestinationPressed()
+{
+	// set flag to keep updating destination until released
+	bMoveToMouseCursor = true;
+}
+
+void AExTDSUEPlayerController::OnSetDestinationReleased()
+{
+	// clear flag to indicate we should stop updating the destination
+	bMoveToMouseCursor = false;
+}
